Round 3

+++ apps/commerce-api/src/main/java/com/loopers/application/point/UserPointFacade.java
@@ -1,6 +1,5 @@
 package com.loopers.application.point;
 
-import com.loopers.domain.point.UserPointModel;
 import com.loopers.domain.point.UserPointService;
 import com.loopers.domain.user.UserService;
 import com.loopers.support.error.CoreException;
@@ -22,7 +21,7 @@ public class UserPointFacade {
         return UserPointInfo.from(userPointService.getUserPoint(userId));
     }
 
-    public UserPointInfo chargeUserPoint(Long userId, Integer amount) {
+    public UserPointInfo chargeUserPoint(Long userId, Long amount) {
         if ( userService.getUserById(userId) == null ) {
             throw new CoreException(ErrorType.NOT_FOUND, "[userId = " + userId + "] 사용자를 찾을 수 없습니다.");
         }
